requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
numpy==1.24.3
supabase==2.0.0
python-multipart==0.0.6
python-dotenv==1.0.0
structlog==23.2.0
httpx>=0.24.0,<0.25.0
Pillow==10.1.0 